#Motivation Currently, c++ & javascript logs on Android are written directly to logcat instead of using the FLog java class. As a result, it is impossible to customize C++ and javascript logging by providing a different LoggerDelegate implementation.
With this change, it is possible to enable saving logs to file by providing a custom implementation of LoggingDelegate.

Without this change, the logs written by JS and C++ code will never land to the file.

#Test Plan
Make sure that logging still works.
Make custom LoggingDelegate which logs to the file.
Make sure that all logs are saved.

#Release Notes
[ANDROID] [ENHANCEMENT] [Logging] - All logs are written using the FLog class. It is possible to customize logging behavior across the board.